+++ src/reduxStore/persistedRootReducer.ts
@@ -12,6 +12,7 @@ import { signedMessageInfoReducer } from 'reduxStore/slices/signedMessageInfoSli
 import toasts from 'reduxStore/slices/toastsSlice';
 import transactionsInfo from 'reduxStore/slices/transactionsInfoSlice';
 import transactions from 'reduxStore/slices/transactionsSlice';
+import batchTransactions from 'reduxStore/slices/batchTransactionsSlice';
 import { ReducersEnum } from 'types/reducers.types';
 
 const config: {
@@ -65,6 +66,10 @@ const transactionsReducerPersistConfig = getSessionStoragePersistConfig(
   'sdk-dapp-transactions',
   [ReducersEnum.transactionsToSign]
 );
+const batchTransactionsReducerPersistConfig = getSessionStoragePersistConfig(
+  'sdk-dapp-batchTransactions',
+  [ReducersEnum.batchTransactions]
+);
 const toastsReducerPersistConfig =
   getSessionStoragePersistConfig('sdk-dapp-toasts');
 const signedMessageInfoersistConfig = getSessionStoragePersistConfig(
@@ -93,6 +98,10 @@ export const sessionStorageReducers = {
   [ReducersEnum.transactionsInfo]: persistReducer(
     transactionsInfoPersistConfig,
     transactionsInfo
+  ),
+  [ReducersEnum.batchTransactions]: persistReducer(
+    batchTransactionsReducerPersistConfig,
+    batchTransactions
   ),
   [ReducersEnum.signedMessageInfo]: persistReducer(
     signedMessageInfoersistConfig,