+++ src/types/transactions.types.ts
@@ -9,6 +9,7 @@ import type {
   TransactionServerStatusesEnum,
   TransactionTypesEnum
 } from './enums.types';
+import { BatchTransactionsResponseType } from './serverTransactions.types';
 
 export interface TransactionsToSignType {
   transactions: IPlainTransactionObject[];
@@ -116,6 +117,7 @@ export interface TransactionsDisplayInfoType {
   submittedMessage?: string;
   transactionDuration?: number;
   timedOutMessage?: string;
+  invalidMessage?: string;
 }
 
 export interface SendSimpleTransactionPropsType {
@@ -206,6 +208,12 @@ export interface CustomTransactionInformation {
 export interface SendTransactionReturnType {
   error?: string;
   sessionId: string | null;
+}
+
+export interface SendBatchTransactionReturnType {
+  error?: string;
+  batchId: string | null;
+  data?: BatchTransactionsResponseType | null;
 }
 
 export type GetTransactionsByHashesType = (