+++ src/utils/parseNavigationParams.ts
@@ -1,4 +1,5 @@
 import { clearNavigationHistory } from './clearNavigationHistory';
+import { isWindowAvailable } from './isWindowAvailable';
 
 interface ParseNavigationParamsOptionsType {
   search?: string;
@@ -6,7 +7,7 @@ interface ParseNavigationParamsOptionsType {
 }
 
 const defaultOptions: ParseNavigationParamsOptionsType = {
-  search: window?.location?.search,
+  search: isWindowAvailable() ? window.location.search : '',
   removeParams: []
 };
 
@@ -20,7 +21,8 @@ export const parseNavigationParams = (
   options = defaultOptions
 ) => {
   let params: Record<string, string> = {};
-  const search = options.search ?? window?.location?.search;
+  const defaultSearch = isWindowAvailable() ? window.location.search : '';
+  const search = options.search ?? defaultSearch;
 
   if (search) {
     const urlSearchParams = search ? new URLSearchParams(search) : [];