+++ src/utils/transactions/batch/getIsSequential.ts
@@ -1,0 +1,9 @@
+import { SignedTransactionType } from 'types';
+
+export function getIsSequential({
+  transactions
+}: {
+  transactions?: SignedTransactionType[] | SignedTransactionType[][];
+}) {
+  return transactions?.some((transaction) => Array.isArray(transaction));
+}