+++ src/utils/transactions/batch/sequentialToFlatArray.ts
@@ -1,0 +1,12 @@
+import { SignedTransactionType } from 'types';
+import { getIsSequential } from './getIsSequential';
+
+export function sequentialToFlatArray({
+  transactions = []
+}: {
+  transactions?: SignedTransactionType[] | SignedTransactionType[][];
+}) {
+  return getIsSequential({ transactions })
+    ? transactions.flat()
+    : (transactions as SignedTransactionType[]);
+}